feat(zkp): tests nearly working
