+++ zkp/__tests__/simple-batch.test.js
@@ -1,12 +1,9 @@
 /* eslint-disable import/no-unresolved */
 
-import utils from 'zkp-utils';
-import { getEthAccounts, getAccountBalance } from '../src/account-utils/account-utils';
-
-import controller from '../src/batch-functions';
-import fcontroller from '../src/f-token-controller';
-import vk from '../src/vk-controller';
-import { FToken } from '../src/contract-abstractions';
+import utils from '../src/zkpUtils';
+import bc from '../src/web3';
+import controller from '../src/f-token-controller';
+import { getVkId, getContract } from '../src/contractUtils';
 
 jest.setTimeout(7200000);
 
@@ -23,12 +20,17 @@ let Z_A_C;
 // storage for z indexes
 let zInd1;
 let zInd2;
-let ftoken;
 let commitments = [];
+let accounts;
+let fTokenShieldJson;
+let fTokenShieldAddress;
 
 beforeAll(async () => {
-  await vk.runController();
-  ftoken = await FToken.deployed(); // get an instance of the ERC20 contract
+  if (!(await bc.isConnected())) await bc.connect();
+  accounts = await (await bc.connection()).eth.getAccounts();
+  const { contractJson, contractInstance } = await getContract('FTokenShield');
+  fTokenShieldAddress = contractInstance.address;
+  fTokenShieldJson = contractJson;
   for (let i = 0; i < PROOF_LENGTH; i++) {
     S_B_E[i] = utils.rndHex(32);
     pkB[i] = utils.ensure0x(utils.zeroMSBs(utils.strip0x(utils.hash(skB)).padStart(32, '0')));
@@ -49,28 +51,32 @@ describe('f-token-controller.js tests', () => {
 
   test('Should create 10000 tokens in accounts[0]', async () => {
     // fund some accounts with FToken
-    const accounts = await getEthAccounts();
     const AMOUNT = 10000;
-    const bal1 = await ftoken.balanceOf.call(accounts[0]);
-    await ftoken.mint(accounts[0], AMOUNT, {
-      from: accounts[0],
-      gas: 4000000,
-    });
-    const bal2 = await ftoken.balanceOf.call(accounts[0]);
+    const bal1 = await controller.getBalance(accounts[0]);
+    await controller.buyFToken(AMOUNT, accounts[0]);
+    const bal2 = await controller.getBalance(accounts[0]);
     expect(AMOUNT).toEqual(bal2 - bal1);
   });
 
   test('Should mint an ERC-20 commitment Z_A_C for Alice of value C', async () => {
-    const accounts = await getEthAccounts();
-    const [zTest, zIndex] = await fcontroller.mint(C, pkA, S_A_C, accounts[0]);
+    const { commitment: zTest, commitmentIndex: zIndex } = await controller.mint(
+      C,
+      pkA,
+      S_A_C,
+      await getVkId('MintCoin'),
+      {
+        account: accounts[0],
+        fTokenShieldJson,
+        fTokenShieldAddress,
+      },
+    );
     zInd1 = parseInt(zIndex, 10);
     expect(Z_A_C).toEqual(zTest);
   });
 
-  test('Should transfer ERC-20 commitments of various values to 19 receipients and get change', async () => {
+  test.skip('Should transfer ERC-20 commitments of various values to 19 receipients and get change', async () => {
     // the E's becomes Bobs'.
-    const accounts = await getEthAccounts();
-    const bal1 = await getAccountBalance(accounts[0]);
+    const bal1 = await controller.getBalance(accounts[0]);
     const response = await controller.simpleFungibleBatchTransfer(
       C,
       E,
@@ -84,7 +90,7 @@ describe('f-token-controller.js tests', () => {
     );
     zInd2 = parseInt(response.z_E_index, 10);
     commitments = response.z_E;
-    const bal2 = await getAccountBalance(accounts[0]);
+    const bal2 = await controller.getBalance(accounts[0]);
     const wei = parseInt(bal1, 10) - parseInt(bal2, 10);
     console.log('gas consumed was', wei / 20e9);
     console.log('approx total cost in USD @$200/ETH was', wei * 200e-18);
@@ -92,13 +98,12 @@ describe('f-token-controller.js tests', () => {
   });
 
   test.skip('Should transfer a pair of the 20 ERC-20 commitments that have just been created', async () => {
-    const accounts = await getEthAccounts();
     const c = '0x00000000000000000000000000000002';
     const d = '0x00000000000000000000000000000002';
     const e = '0x00000000000000000000000000000001';
     const f = '0x00000000000000000000000000000003';
     const pkE = await utils.rndHex(32); // public key of Eve, who we transfer to
-    await fcontroller.transfer(
+    await controller.transfer(
       c,
       d,
       e,
