+++ zkp/contracts/FTokenShield.sol
@@ -28,16 +28,16 @@ contract FTokenShield is Ownable {
          15  16  17 18  19  20  21  22  23  24  25  26  27  28  29  30
 
 depth row  width  st#     end#
-  1    0   2^0=1  w=0   2^1-2=0
-  2    1   2^1=2  w=1   2^2-2=2
-  3    2   2^2=4  w=3   2^3-2=6
-  4    3   2^3=8  w=7   2^4-2=14
-  5    4   2^4=16 w=15  2^5-2=30
+  1    0   2**0=1  w=0   2**1-2=0
+  2    1   2**1=2  w=1   2**2-2=2
+  3    2   2**2=4  w=3   2**3-2=6
+  4    3   2**3=8  w=7   2**4-2=14
+  5    4   2**4=16 w=15  2**5-2=30
 
   d = depth = 5
   r = row number
-  w = width = 2^(depth-1) = 2^3 = 16
-  #nodes = (2^depth)-1 = 2^5-2 = 30
+  w = width = 2**(depth-1) = 2**3 = 16
+  #nodes = (2**depth)-1 = 2**5-2 = 30
 
   */
